[Enhancement #368] Refactor Users table to support filtering, paging and sorting.
Using react-table. Extract common table infrastructure into a component.
