+++ src/component/administration/user/UserRoles.tsx
@@ -1,9 +1,9 @@
 import { connect } from "react-redux";
-import Utils from "../../util/Utils";
-import { getLocalized } from "../../model/MultilingualString";
-import UserRole, { UserRoleData } from "../../model/UserRole";
-import User from "../../model/User";
-import TermItState from "../../model/TermItState";
+import Utils from "../../../util/Utils";
+import { getLocalized } from "../../../model/MultilingualString";
+import UserRole, { UserRoleData } from "../../../model/UserRole";
+import User from "../../../model/User";
+import TermItState from "../../../model/TermItState";
 
 export function filterActualRoles(types: string[], roles: UserRoleData[]) {
   return Utils.sanitizeArray(roles).filter((r) => types.indexOf(r.iri) >= 0);
