+++ src/component/administration/user/UserRolesEdit.tsx
@@ -8,15 +8,15 @@ import {
   ModalBody,
   ModalHeader,
 } from "reactstrap";
-import UserRole, { UserRoleData } from "../../model/UserRole";
-import { getLocalized } from "../../model/MultilingualString";
-import Select from "../misc/Select";
+import UserRole, { UserRoleData } from "../../../model/UserRole";
+import { getLocalized } from "../../../model/MultilingualString";
+import Select from "../../misc/Select";
 import { filterActualRoles } from "./UserRoles";
-import User from "../../model/User";
-import TermItState from "../../model/TermItState";
-import { useI18n } from "../hook/useI18n";
-import Utils from "../../util/Utils";
-import VocabularyUtils from "../../util/VocabularyUtils";
+import User from "../../../model/User";
+import TermItState from "../../../model/TermItState";
+import { useI18n } from "../../hook/useI18n";
+import Utils from "../../../util/Utils";
+import VocabularyUtils from "../../../util/VocabularyUtils";
 
 interface UserRolesEditProps {
   user: User;
